# TASK: improve performance of `getRelativeNodePathByUriPathSegmentProperties`