Use `NodeSearchService` to query for nodes with given uriPathSegment.

On pages with a lot of nodes on the same level (>1000) this drastically
improves performance.