@dimaip 2016-08-26
oups, forgot to adjust the tests, will do it soonish :-1: