@dimaip 2016-08-27
I give up on adjusting unit tests for this one.
My opinion: it doesn't make sense to unit test such huge and messy class as FrontentNodeRoutePartHandler, especially after this change, where part of essential logic is moved into the search service. I would propose to drop unit tests for it altogether until it's rewritten and just improve functional tests coverage for now (if needed).

Do you have better ideas?
Please help!
/cc @kdambekalns @kitsunet