@kitsunet 2016-08-27
First of all the syntax error should be fixed ;)
I am not a big fan of our tests but it's all we have, especially if we want to refactor, so 👎 for removing tests.