@dimaip 2016-08-27
@kitsunet oups, side-effects of night coding...
OK, I will try to spend some time to understand how to refactor unit tests to work with this change. The problem here is that I have replaced the crucial part of this class by for external service, and it would not be easy to mock it in the right way for me. Guess it's a good place to learn.