@dimaip 2016-08-29
@kitsunet @kdambekalns okay guys, it wasn't as scary as I thought, I think I have fixed the tests. Please give it a look when you have time.