@hlubek 2016-08-30
@dimaip Same here, it would be nice to get this into 2.3 (if possible)