@kitsunet 2016-08-30
Would be nice, but I am not really confident to merge this into 2.3 yet.