@dimaip 2016-08-31
I'm testing performance now and noticed a really weird thing: if I take the current master, without applying my changes, it works really fast in the live workspace (0.18s for resolving a node with 7000 siblings), but it slows to a crawl when testing from the BE (3.48s). Can anyone make sense out of why it happens?