@dimaip 2016-08-31
Ah so I see that `getChildNodes` has its own layer of caching: `$this->context->getFirstLevelNodeCache()->getChildNodesByPathAndNodeTypeFilter`
So it's super fast once it's cached.