@dimaip 2016-08-31
Back to @aertmann's hint. Here are the performance metrics (all measured in backend):

| Metric type | master | this PR |
|----------------|-----------|-----------|
| 10 nesting levels | 0.025s | **1.13s** |
| 7000 siblings | **3.78s** | 0.22s |