@dimaip 2016-09-02
Anyways,  I've AOPed myself a fix tailored for my current project: https://github.com/psmb/PsmbDistr/commit/33375aa8a60db43574f30416f87a54c93b0878ce

I will pause project on this PR until somebody tells me which way to go. I'm stuck and have no idea how to fix it universally.